Squashed commit of the following:
commit 6a882096bdc04c39f6afea76d1224f33d0366eec
Author: vlaya8 <vlaya8@protonmail.com>
Date:   Sun Jul 5 03:10:42 2020 +0200

    Replace PermissionDenied which could be caused by certain links

commit 00192dd2894183e25027c0096590968a683af836
Author: vlaya8 <vlaya8@protonmail.com>
Date:   Sun Jul 5 02:44:46 2020 +0200

    Add personal group link and group visibility info to profile

commit 275bfda7a8f17d2a22447c1dff5ecd223485df7f
Author: vlaya8 <vlaya8@protonmail.com>
Date:   Sun Jul 5 02:31:52 2020 +0200

    Prevent writer from editing a group's info

commit 5ad1b0fff688b291d66c7352f796d915ac837861
Author: vlaya8 <vlaya8@protonmail.com>
Date:   Sun Jul 5 02:30:50 2020 +0200

    Delete obsolete public_group and listed_group fields

commit 59811a7dc61b9dcb5cbdd80750902f20ee648931
Author: vlaya8 <vlaya8@protonmail.com>
Date:   Sun Jul 5 02:30:05 2020 +0200

    Use group_visibility field intead of obsolete public_group and listed_group

commit f0d4032a85f963775bcbe0e7c2b9c7c4b3aa611c
Author: vlaya8 <vlaya8@protonmail.com>
Date:   Sun Jul 5 02:23:03 2020 +0200

    Change GROUP_STATUS_CHOICES to GROUP_VISIBILITY_CHOICES

commit bb28da6be11f7ae97c397907174896dd15365010
Author: vlaya8 <vlaya8@protonmail.com>
Date:   Sun Jul 5 02:21:59 2020 +0200

    Change group_status field to group_visibility

commit 20487ae5e3402b805ede87f077dd4008fed82f56
Author: vlaya8 <vlaya8@protonmail.com>
Date:   Sun Jul 5 02:20:15 2020 +0200

    Add group visibility info

commit 4ea22e75bb9c81d374b4a7c598053c8d7bd1f784
Author: vlaya8 <vlaya8@protonmail.com>
Date:   Sun Jul 5 02:14:45 2020 +0200

    Fix bug in upsert group

commit a596fb7cd14048291a4acfbf7f8abcad97383fae
Author: vlaya8 <vlaya8@protonmail.com>
Date:   Sun Jul 5 01:04:37 2020 +0200

    Adapt HTML to avoid using custom widget

commit 66a146a0de2e355448fb85fe84ceb521dd270ce2
Author: vlaya8 <vlaya8@protonmail.com>
Date:   Sun Jul 5 00:45:16 2020 +0200

    Change select widget to pure CSS select

commit ff850cb0aefbb67543d07b6d7e9f450d5f50958d
Author: vlaya8 <vlaya8@protonmail.com>
Date:   Sat Jul 4 23:05:08 2020 +0200

    Add group_status field

commit a17bdad53ce69915ff09a97bd4d184da192bc413
Author: vlaya8 <vlaya8@protonmail.com>
Date:   Sat Jul 4 22:35:43 2020 +0200

    Add group type information and prevent inviting members in personal group

commit 981e5e3bc26afd5b7502195061395a89d40e7309
Author: vlaya8 <vlaya8@protonmail.com>
Date:   Sat Jul 4 22:32:58 2020 +0200

    Allow user to edit their personal group preferences

commit 1e4b0b4a2ac77eb695a9320506b520f195a1b437
Merge: 7637cac 33575b8
Author: vlaya8 <vlaya8@protonmail.com>
Date:   Sat Jul 4 21:45:58 2020 +0200

    Merge branch 'dev' into groups

commit 7637cac764e76e7158adfd9f79052a5af0b9f7a7
Merge: 9c62bf5 3a0e8a3
Author: vlaya8 <vlaya8@protonmail.com>
Date:   Thu Jul 2 22:32:29 2020 +0200

    Merge branch 'dev' into groups

commit 9c62bf5bfbb9b930d6813aa6711b623874d39de9
Author: vlaya8 <vlaya8@protonmail.com>
Date:   Thu Jul 2 22:31:49 2020 +0200

    Prevent user from deleting last member of a group

commit c66186f4edc442c4dafa25dd41f1ba42934add12
Author: vlaya8 <vlaya8@protonmail.com>
Date:   Thu Jul 2 22:05:50 2020 +0200

    Add link to member and delete button in view_group

commit 5b1aa385decc31e9dd91851a3c2e6fd230981898
Author: vlaya8 <vlaya8@protonmail.com>
Date:   Thu Jul 2 22:01:28 2020 +0200

    Restructure view_group html

+++ tags/views/views.py
@@ -22,7 +22,7 @@ from notifications.models import Notification
 from ..exceptions import UserError
 from ..models import Entry,Tag,Tree
 from ..models import TreeUserGroup, Role, Member
-from ..forms import EntryForm, TreeForm, GroupForm, MemberInvitationForm
+from ..forms import EntryForm, TreeForm, GroupForm, MemberInvitationForm, ProfileForm
 from .utilities import *
 
 import json
@@ -221,16 +221,68 @@ class ProfileView(UserDataView):
 
     template_name = 'tags/registration/profile.html'
 
-    def get_context_data(self, request, **kwargs):
+    def get_context_data(self, request, form=None, **kwargs):
 
         context = super().get_context_data(request, **kwargs)
 
         has_edit_permission = (self.user.username == self.link_user.username)
+        if has_edit_permission:
+
+            group = self.user.get_user_group()
+            if form == None:
+                form = ProfileForm(initial={
+                                             'username': self.user.username,
+                                             'personal_group_visibility': group.group_visibility,
+                                           })
+            context.update({'form': form})
+
+        link_group = self.link_user.get_user_group()
+
+        if link_group.group_visibility == TreeUserGroup.PRIVATE:
+            group_visibility = "Private"
+        elif link_group.group_visibility == TreeUserGroup.PUBLIC:
+            group_visibility = "Unlisted and Public"
+        else:
+            group_visibility = "Listed and Public"
 
         context.update({
                          'has_edit_permission': has_edit_permission,
-                      })
+                         'group_visibility': group_visibility,
+                         'has_group_view_permission': link_group.is_public(),
+                       })
+
         return context
+
+    def post_return(self, request, **kwargs):
+
+        if 'edit_profile' in request.POST:
+
+            form = ProfileForm(request.POST)
+
+            if form.is_valid():
+
+                if self.user.username != self.link_user.username:
+                    raise PermissionDenied("You don't have permission to edit this profile")
+
+                username = form.cleaned_data['username']
+                personal_group_visibility = form.cleaned_data['personal_group_visibility']
+
+                group = self.user.get_user_group()
+
+                group.group_visibility = personal_group_visibility
+                group.name = username
+                group.save()
+
+                self.user.username = username
+                self.user.save()
+
+
+                self.redirect_url = 'tags:profile'
+                self.redirect_kwargs['username'] = self.user.username
+            else:
+                context = self.get_context_data(request, form=form, **kwargs)
+                return self.get_return(request, context, **kwargs)
+        return super().post_return(request, **kwargs)
 
 class ChangePasswordView(auth_views.PasswordChangeView):
 
@@ -249,9 +301,9 @@ class ManageGroupsView(UserDataView):
         saved_group_list = self.user.get_saved_groups()
         listed_group_list = TreeUserGroup.get_listed_groups()
 
-        joined_groups = [(group, self.user.has_group_admin_permission(group)) for group in joined_group_list]
-        saved_groups = [(group, self.user.has_group_admin_permission(group)) for group in saved_group_list]
-        listed_groups = [(group, self.user.has_group_admin_permission(group)) for group in listed_group_list]
+        joined_groups = [(group, not group.single_member and self.user.has_group_admin_permission(group)) for group in joined_group_list]
+        saved_groups = [(group, not group.single_member and self.user.has_group_admin_permission(group)) for group in saved_group_list]
+        listed_groups = [(group, not group.single_member and self.user.has_group_admin_permission(group)) for group in listed_group_list]
 
         context.update({
                          'joined_groups': joined_groups,
@@ -270,8 +322,13 @@ class ManageGroupsView(UserDataView):
             group_id = int(request.POST['delete_group_id'])
 
             group = get_object_or_404(TreeUserGroup, pk=group_id)
+
             if not self.user.has_group_writer_permission(group):
                 raise PermissionDenied("You don't have permission to delete this group")
+
+            if group.single_member:
+                raise PermissionDenied("You don't have permission to delete this group")
+
             group.delete()
 
         self.redirect_url = 'tags:manage_groups'
@@ -293,7 +350,7 @@ class UpsertGroupView(UserDataView):
             if not self.user.has_group_writer_permission(group):
                 raise PermissionDenied("You don't have permission to edit this group")
 
-            data = {"name": group.name, "group_id": group_id, "public_group": group.public_group, "listed_group": group.listed_group}
+            data = {"name": group.name, "group_visibility": group.group_visibility, "group_id": group_id}
         # If user wants to add a group
         else:
             data = {"group_id": SpecialID['NEW_ID']}
@@ -306,6 +363,45 @@ class UpsertGroupView(UserDataView):
 
         return context
 
+    def post_return(self, request, **kwargs):
+
+        # Process a group which got added or edited
+        if 'upsert_group' in request.POST:
+
+            form = GroupForm(request.POST)
+
+            if form.is_valid():
+
+                group_name = form.cleaned_data['name']
+                group_visibility = form.cleaned_data['group_visibility']
+                group_id = form.cleaned_data['group_id']
+
+                # If the group has been edited
+                if group_id != SpecialID['NEW_ID']:
+                    group = get_object_or_404(TreeUserGroup, pk=group_id)
+
+                    if not self.user.has_group_writer_permission(group):
+                        raise PermissionDenied("You don't have permission to edit this group")
+
+                    group.name = group_name
+                    group.group_visibility = group_visibility
+                else:
+                    group = TreeUserGroup.objects.create(name=group_name, single_member=False, group_visibility=group_visibility)
+                    admin_role = Role.objects.filter(name="admin").first()
+                    member = Member.objects.create(user=self.user, role=admin_role, group=group)
+
+                group.save()
+
+                self.redirect_url = 'tags:view_group'
+                self.redirect_kwargs['group_id'] = group.id
+                self.redirect_kwargs['username'] = self.user.username
+            else:
+                context = self.get_context_data(request, group_id, invite_form=form, **kwargs)
+                return self.get_return(request, context, **kwargs)
+
+        return super().post_return(request, **kwargs)
+
+
 class ViewGroupView(UserDataView):
 
     template_name = 'tags/view_group.html'
@@ -317,57 +413,34 @@ class ViewGroupView(UserDataView):
         group = get_object_or_404(TreeUserGroup, pk=group_id)
 
         if not self.user.has_group_reader_permission(group):
-            raise PermissionDenied("You don't have permission to view that group")
+            context.update({'has_reader_permission': False})
+            return context
 
         members = group.member_set.all()
 
         if invite_form == None:
             invite_form = MemberInvitationForm(group, initial={})
 
+        if group.group_visibility == TreeUserGroup.PRIVATE:
+            group_visibility = "Private"
+        elif group.group_visibility == TreeUserGroup.PUBLIC:
+            group_visibility = "Unlisted and Public"
+        else:
+            group_visibility = "Listed and Public"
+
         context.update({
                          'group': group,
+                         'has_reader_permission': True,
                          'has_writer_permission': self.user.has_group_writer_permission(group),
                          'has_admin_permission': self.user.has_group_admin_permission(group),
                          'members': members,
                          'invite_form': invite_form,
+                         'group_visibility': group_visibility,
                       })
 
         return context
 
     def process_post(self, request, group_id=SpecialID['NEW_ID'], **kwargs):
-
-        # Process a group which got added or edited
-        if 'upsert_group' in request.POST:
-
-            form = GroupForm(request.POST)
-
-            if form.is_valid():
-
-                group_name = form.cleaned_data['name']
-                group_id = form.cleaned_data['group_id']
-                public_group = form.cleaned_data['public_group']
-                listed_group = form.cleaned_data['listed_group']
-
-                # If the group has been edited
-                if group_id != SpecialID['NEW_ID']:
-                    group = get_object_or_404(TreeUserGroup, pk=group_id)
-
-                    if not self.user.has_group_writer_permission(group):
-                        raise PermissionDenied("You don't have permission to edit this group")
-
-                    group.name = group_name
-                    group.public_group = public_group
-                    group.listed_group = listed_group
-                else:
-                    group = TreeUserGroup.objects.create(name=group_name, single_member=False, public_group=public_group, listed_group=listed_group)
-                    admin_role = Role.objects.filter(name="admin").first()
-                    member = Member.objects.create(user=self.user, role=admin_role, group=group)
-
-                group.save()
-
-                self.redirect_url = 'tags:view_group'
-                self.redirect_kwargs['group_id'] = group.id
-                self.redirect_kwargs['username'] = self.user.username
 
         # If a member has been deleted
         if 'delete_member_id' in request.POST:
@@ -425,9 +498,6 @@ class TreeView(UserDataView):
         # Setup group argument
         self.group = get_object_or_404(TreeUserGroup, name=group_name)
 
-        if not self.user.has_group_reader_permission(self.group):
-            raise PermissionDenied("You don't have permission to view this group")
-
         # Setup tree argument
         if tree_id == SpecialID['DEFAULT_ID']:
             self.current_tree = Tree.objects.filter(group__id=self.group.id).first()
@@ -441,6 +511,10 @@ class TreeView(UserDataView):
     def get_context_data(self, request, **kwargs):
 
         context = super().get_context_data(request, **kwargs)
+
+        if not self.user.has_group_reader_permission(self.group):
+            context.update({'has_reader_permission': False})
+            return context
 
         tree_list = []
         tree_add_form = TreeForm(initial={'tree_id': SpecialID['NEW_ID'], 'delete_tree': False})
@@ -460,6 +534,7 @@ class TreeView(UserDataView):
             current_tree_id = self.current_tree.id
 
         context.update({
+                    'has_reader_permission': True,
                     'tree_list': tree_list,
                     'tree_add_form': tree_add_form,
                     'tree_bar_data': json.dumps({'nb_trees': len(tree_list), 'tree_ids': tree_ids}),
@@ -476,6 +551,9 @@ class TreeView(UserDataView):
         # Default redirect
         self.redirect_kwargs['group_name'] = self.group.name
         self.redirect_url = 'tags:view_tree'
+
+        if not self.user.has_group_reader_permission(self.group):
+            raise UserError("")
 
         if 'tree_form' in request.POST:
             if not self.user.has_group_writer_permission(self.group):
